+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Manifest Version 1.4.12
+
+- added more checks to release script
+- added network icons and templates for the following networks: World Chain, Linea, Scroll
+
 ## Manifest Version 1.4.11
 
 - bumped all dependencies to the latest versions